move testdatabasemutationbuilder
